Fix: Update all migrations to use unsignedBigInteger for foreign keys
- Changed all integer mahasiswa_id, dosen_id foreign keys to unsignedBigInteger
- Fixed foreign key constraint errors in multiple migrations
- Successfully completed all pending migrations
- Tugas table and related tables now created successfully
- Fixed attendance_permits, gamification, kas_voting, academic tables
- All database tables now properly set up for dosen functionality

+++ database/migrations/2026_01_13_000001_create_kas_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('kas', function (Blueprint $table) {
             $table->id();
-            $table->integer('mahasiswa_id')->nullable(); // null for expense
+            $table->unsignedBigInteger('mahasiswa_id')->nullable(); // null for expense
             $table->enum('type', ['income', 'expense'])->default('income');
             $table->decimal('amount', 12, 2);
             $table->string('description')->nullable();
